requirements.txt
ollama>=0.1.5
pytesseract>=0.3.10
Pillow>=10.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
streamlit
openai